# docs/swagger.yaml
basePath: /api/v1
definitions:
  dtos.AddExchangeReq:
    properties:
      name:
        description: Binance
        type: string
      ws_url:
        description: wss://ws-api.binance.com:443/ws-api/v3
        type: string
    type: object
  dtos.AddExchangeRes:
    properties:
      id:
        description: "1"
        type: string
      name:
        description: Binance
        type: string
      ws_url:
        description: wss://ws-api.binance.com:443/ws-api/v3
        type: string
    type: object
  dtos.AddSignalIntervalReq:
    properties:
      exchange_id:
        description: "1"
        type: string
      interval:
        description: 1m, 5m, 15m, 1h, 4h, 1d
        type: string
      symbol:
        description: BTCUSDT
        type: string
    type: object
  dtos.AddSymbolReq:
    properties:
      exchange_id:
        type: string
      symbol:
        description: BTCUSDT
        type: string
    type: object
  dtos.GetExchangeRes:
    properties:
      id:
        type: string
      name:
        description: Binance
        type: string
      ws_url:
        description: wss://ws-api.binance.com:443/ws-api/v3
        type: string
    type: object
  dtos.GetSymbolRes:
    properties:
      exchange_id:
        type: string
      id:
        type: string
      is_active:
        description: 1 active, 2 passive
        type: integer
      symbol:
        type: string
    type: object
  dtos.UpdateExchangeReq:
    properties:
      id:
        type: string
      name:
        description: Binance
        type: string
      ws_url:
        description: wss://ws-api.binance.com:443/ws-api/v3
        type: string
    type: object
  dtos.UpdateExchangeRes:
    properties:
      id:
        type: string
      name:
        description: Binance
        type: string
      ws_url:
        description: wss://ws-api.binance.com:443/ws-api/v3
        type: string
    type: object
  dtos.UpdateSignalIntervalReq:
    properties:
      exchange_id:
        description: "1"
        type: string
      id:
        description: "1"
        type: string
      interval:
        description: 1m, 5m, 15m, 1h, 4h, 1d
        type: string
      symbol:
        description: BTCUSDT
        type: string
    type: object
  dtos.UpdateSymbolReq:
    properties:
      exchange_id:
        type: string
      id:
        type: string
      symbol:
        type: string
    type: object
host: localhost:8001
info:
  contact: {}
  description: Crypto Trade API Documentation
  title: Crypto Trade API
  version: "1.0"
paths:
  /:
    post:
      consumes:
      - application/json
      description: Add Symbol Route
      parameters:
      - description: Add Symbol Request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dtos.AddSymbolReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add Symbol
      tags:
      - Symbol Endpoints
  /exchanges:
    get:
      consumes:
      - application/json
      description: Retrieves all exchanges
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.GetExchangeRes'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Get all exchanges
      tags:
      - exchanges
    post:
      consumes:
      - application/json
      description: Add a new exchange
      parameters:
      - description: Exchange information
        in: body
        name: exchange
        required: true
        schema:
          $ref: '#/definitions/dtos.AddExchangeReq'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created exchange
          schema:
            $ref: '#/definitions/dtos.AddExchangeRes'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Add exchange
      tags:
      - exchanges
  /exchanges/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an exchange by its ID
      parameters:
      - description: Exchange ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Delete an exchange
      tags:
      - exchanges
    get:
      consumes:
      - application/json
      description: Retrieves an exchange by its ID
      parameters:
      - description: Exchange ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.GetExchangeRes'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Get exchange by ID
      tags:
      - exchanges
    put:
      consumes:
      - application/json
      description: Updates an existing exchange with the provided information
      parameters:
      - description: Exchange ID
        in: path
        name: id
        required: true
        type: string
      - description: Exchange update information
        in: body
        name: exchange
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateExchangeReq'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated exchange
          schema:
            $ref: '#/definitions/dtos.UpdateExchangeRes'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Update an exchange
      tags:
      - exchanges
  /signal/interval:
    get:
      consumes:
      - application/json
      description: Get All Signal Intervals Route
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get All Signal Intervals
      tags:
      - Signal Endpoints
    post:
      consumes:
      - application/json
      description: Add Signal Interval Route
      parameters:
      - description: Add Symbol Request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dtos.AddSignalIntervalReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add Signal Interval
      tags:
      - Signal Endpoints
  /signal/interval/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Signal Interval Route
      parameters:
      - description: Signal ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete Signal Interval
      tags:
      - Signal Endpoints
    get:
      consumes:
      - application/json
      description: Get Signal Interval by ID Route
      parameters:
      - description: Signal ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Signal Interval by ID
      tags:
      - Signal Endpoints
    put:
      consumes:
      - application/json
      description: Update Signal Interval Route
      parameters:
      - description: Signal ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Signal Request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateSignalIntervalReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update Signal Interval
      tags:
      - Signal Endpoints
  /symbol:
    get:
      consumes:
      - application/json
      description: Get All Symbols Route
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.GetSymbolRes'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get All Symbols
      tags:
      - Symbol Endpoints
  /symbol/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Symbol Route
      parameters:
      - description: Symbol ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete Symbol
      tags:
      - Symbol Endpoints
    get:
      consumes:
      - application/json
      description: Get Symbol by ID Route
      parameters:
      - description: Symbol ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.GetSymbolRes'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Symbol by ID
      tags:
      - Symbol Endpoints
    put:
      consumes:
      - application/json
      description: Update Symbol Route
      parameters:
      - description: Symbol ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Symbol Request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateSymbolReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update Symbol
      tags:
      - Symbol Endpoints
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"